FlowTrigger and FlowTriggerDependency update (#1609)
This PR made several changes to FlowTrigger/FlowTriggerDependency class

Made FlowTrigger and FlowTriggerDependency Serializable. This is to required by Quartz scheduler to persist scheduled job.

Added error message when doing argument validation. This is for better user experience when misconfigured flow trigger is uploaded.

Converted dependencies type from list to a hashmap with name as key for convenient lookup.