adding job logger to credential instance (#1600)
In order to allow users to debug credential issues easily, this PR proposes adding job logger into the credential instance, so that related log messages will be shown in Azkaban web UI.

Notice that this PR is backward compatible. In the credential client code side, we have both 2-parameter and 3-parameter constructor. If we need to roll back this commit, it will not fail.