Add boolean property for evicting executions
New boolean property "azkaban.evictExecutions". Default: false.

This can be used to decide whether Azkaban should ever give up on getting a response from executor.

As before, an eventual successful response may still indicate that an execution is not running on that executor any more, which will cause cleaning it up as usual.