fix reportal regression bug introduced by security fix #1325 (#1730)
We found a security hole in #1325 that any job is able to leverage execute-as-user to impersonate any unix account. when we tryied rolling out this fix to Reportal server, we run into an exception for all ReportalPig jobs:

The cause is that the job process is not able to create a file under the same source script folder, because of the change (#1325).

In this PR, I used the most straightforward solution to create the bak file under the root working directory.