job name should be customized for different types of quartz jobs (#1756)
A quartz job is identified by job name and group name. Group name is the combination of project id and flow id. But job name was a constant string. This PR replaces it with dynamic job name variable.
For flow trigger, the job name is FLOW_TRIGGER. This is to provide flexibility in case where same flow could have multiple scheduled quartz jobs in the future.