Refactor DagService shutdown method and add tests (#1765)
* Refactor DagService shutdown method and add tests

Part of #1486

Create a new utility class to encapsulate the generic shutdown logic to
make it easier to test and allow better code reuse.

Change the shutdown API to take a Duration type instead of a long. This is more precise and flexible. 

Also reduced the shutdown timeout from 60 seconds to 10 seconds.

After this change, the dag package will have 100% line
coverage.