Add user.to.proxy validation before reportal uploads the user zip to AZ (#1722)
A user is able to user.to.proxy any user in Reportal today. This has been a security vulnerability issue for long time. This PR proposes a fix to validate user.to.proxy configuration before the zip being uploaded to AZ. If the session user is not part of userToProxy user, error is returned to browser to prevent users saving the Reportal.