Make testCancelAfterJobProcessCreation more reliable
Issue:

The test failed today on trunk

azkaban.jobExecutor.ProcessJobTest > testCancelAfterJobProcessCreation FAILED
    org.junit.ComparisonFailure: expected:<[fals]e> but was:<[tru]e>
        at sun.reflect.NativeConstructorAccessorImpl.newInstance0(Native Method)
        at sun.reflect.NativeConstructorAccessorImpl.newInstance(NativeConstructorAccessorImpl.java:62)
        at sun.reflect.DelegatingConstructorAccessorImpl.newInstance(DelegatingConstructorAccessorImpl.java:45)
        at azkaban.jobExecutor.ProcessJobTest.testCancelAfterJobProcessCreation(ProcessJobTest.java:228)

Cause:

I suspect it is because of the travis CI machine being too busy to cancel
the job before it finishes.

Solution:

Increase the running time of the job from 1 second to 5 seconds.
This will not increase the test running time when the test is successful
since the job will be canceled as soon as possible.
However it will increase the test running time if the cancel logic
doesn't interrupt the job as expected.

Also removed the printout of the stacktrace from the job. This information
doesn't deliver much value and only makes the log noisier.