adding text to indicate that columns are sortable (#1795)
As mentioned in issue #1778, having a tooltip on the table would let the user know that it can be sorted when the header is clicked. This pull request will add the tooltip on scheduling page and history page which are being mostly used.

Sample screenshot:
image

This is a duplicate PR (#1780) but from a different branch, I moved these changes from master->sort-text in my repo.