Add a test to run a v2 flow file with the new DAG engine
This is a step towards building an integration test for the new DAG engine.

Refactored the DagBuilder API to make it easier to use.
Instead of using a DagBuilder class to link nodes, use the name of the
nodes directly.

This facility can also be used in the future to build the tools to run
flows locally for testing purposes.