Fix NPE in FetchActiveFlowDao
Ignore active executions with flow_data=null.

Cleaning them up manually would require too many changes, because ExecutableFlow object is used in many places.

Note that it's still possible to end up having this kind of rows in the DB if the application fails between creation of an execution id in `execution_flows` and uploading the data. I'm not sure if doing both inside a transaction is even possible, so I rather just avoid the NPE and log a warning about this corner case that should be rare.