Fix NPE in FetchActiveFlowDao (#1862)
Ignore active executions with flow_data=null.

For error details, see #1833 (comment).

Cleaning these up manually would require too many changes, because ExecutableFlow object is used in many places.

Note that it's still possible to end up having this kind of rows in the DB if the application fails between creation of an execution id in execution_flows and uploading the data. See here:

azkaban/azkaban-common/src/main/java/azkaban/executor/ExecutionFlowDao.java

Lines 71 to 74 in 2be81eb

 final long id = this.dbOperator.transaction(insertAndGetLastID); 
 logger.info("Flow given " + flow.getFlowId() + " given id " + id); 
 flow.setExecutionId((int) id); 
 updateExecutableFlow(flow); 
I'm not sure if doing both inside a transaction is even possible, so I rather just avoid the NPE and log a warning about this corner case that should be rare.

This weakness was in place already before my change #1833. But before my change it could have only caused an NPE if the same executor id is still used after missing to update the flow data (meaning, using a fixed executor id, or starting up an executor after a sudden crash where the executor didn't get to remove itself from the DB as a part of the shutdown hook).