Log about loadRunningFlows in azkaban-web startup
For example setupExecutors() logs a similar message. However fetching active executors is quick.

Loading the active flows is what happens after it, and it may take a while (last time in my case it took ~20s). It's comforting to see in the log what we're waiting for, when the server is starting.