No recursion in dispatch error handling
The code had a recursion loop like this:

selectExecutorAndDispatchFlow
  -> handleDispatchExceptionCase
    -> selectExecutorAndDispatchFlow
      -> ..

..with max depth of activeExecutors.size()

Iterative approach is easier to reason about, and is totally safe against StackOverflowError.

This is also preliminary refactoring to eventually allow configuring higher max for dispatching errors. Currently, however, max dispatch attempts is capped by the total number of active executors (because each executor is only tried once). In the future I'd like to remove that limit and rely solely on the MAX_DISPATCHING_ERRORS_PERMITTED property.