check if project is removed while uploading (#1916)
We found that upload API doesn't return an error when a deleted project zip is uploaded.

When a project is deleted, Azkaban just marks it to be inactive today. In this PR, we fix this issue by adding the check. 410 Gone client error response code indicates that access to the target resource is no longer available at the origin server and that this condition is likely to be permanent.