Fix flow log CSS (#1883)
The Flow Log positioning was broken

This happens if the bar above (the one that has Submit User, Duration etc.) spans to multiple rows. Even one of them is enough to cause overflowing positioning, but the screenshot above demonstrates an extreme case where the tab is barely visible any more.

As a quick fix, I copied the same rules that are set for div#graphView.container-fill (you can find it right above the changed lines).

the fix sets a fixed height to the log textbox, and as the screenshot shows it gets cut off if the window size is not as high. There must be a way to make the position relative properly.. If that is found, why not apply that better rule also for div#graphView.container-fill?