Make pageSize configurable (#1797)
This will address the issue #1796. A new property called azkaban.display.numOfExecutions will be added in the azkaban.properties and be used to determine the number of rows to be displayed.

As it will default to 16 when the property does not exist, it will not have any impact on the existing page size.