Reportal tableau jobtype (#1932)
This PR adds tableau jobtype for reportal.
Tableau is not real-time: whenever users want to incorporate the newest data into their visualization, they must refresh the extracted data. This tells Tableau to get and show whatever new data has come into the data sources. It's natural for users to want to automatically trigger Tableau refreshes when their jobs are done and their data sources are full of new data. This job is designed to do exactly that - refreshing a Tableau workbook whenever run so users don't have to do so manually.

Since tableau workbook refresh doesn't require authentication, all users have to do is identify the name of the workbook they'd like to refresh and it will do so.