Reportal presto jobtype (#1929)
This PR introduces presto jobtype for reportal. User is able to run presto query through reportal UI with a shared presto account defined in jobtype plugin config. Presto account password could be encrypted and kept in the config file, and it requires a key path file which could be defined in jobtype plugin config to decrypt.