Don't delete (clean) project versions with running executions
When a new version of a project is uploaded, the old versions are cleaned up (zip files are deleted from the DB).

This way the project versions with running executions are skipped in the old versions cleanup.

Even those versions will be eventually cleaned up once those executions have finished, because the basic filter is `version < ?`