Shut down single server if startup fails (#2002)
The single server first starts executor server, then web server.

If web server startup failed, the single server process was left running, making it harder to realize that there was an issue in startup.

This PR makes it easier to spot a startup error by stopping JVM in case of a startup error.
Without this change, the ERROR from web startup is buried in an earlier phase of the logs. The JVM keeps running despite an exception thrown by web start call because executor server's thread remains alive. As the executor server has already successfully started, it will keep logging some periodical stuff into the log file, making it harder to find out what was the last thing that happened on the web server's side.