Unify jvm memory settings validation & improve upload error message (#2111)
If there's an empty value like `"job.max.Xmx="` in some properties, upload fails with this error:

> Installation Failed.
> For input string: ""

So it doesn't help much because it doesn't even tell the name of the problematic property.

This PR improves it by returning a better error message ie. one that includes the property name.

Currently user would need to grab the stack trace, get azkaban source code, and find the line, to know which property caused the upload failure.