Implement "flowPriority" (at the time of polling) feature for new Dispatching Logic (#2140)
* Implement "flowPriority" (at the time of polling) feature for new dispatching logic

"flowPriority" is one of the execution options of a flow, which only takes effect if set by an Azkaban admin. By default all flows have a priority of 5 but if a bigger number is specified from the UI that flow will be dispatched for execution first.

With the new Dispatching Logic setting a high "flowPriority" means that flow will be polled first by executors. 
Note that it has been said "polled first" not "executed first". On the executor server there is a queue to hold executions when the maximum number of threads availables/setup to run flows is reached. This queue is NOT a priority queue.

To implement this feature for the new Dispatching Logic, a new column "flow_priority" needs to be added to table "execution_flows" to speed up access to this info (currently flowPriority is stored in "flow_data" longblob column). When new executions are inserted into "execution_flows" table, user-specified or default flow priority will be inserted as well. Later, executors will poll executions sorting by flow_priority DESC, submit_time ASC, exec_id ASC

* Address review comments

* Set flow_priority column to TINYINT type.

* Add unit tests

* Set submitTime and status attributes outside uploadExecutableFlow method