FlowRunnerYamlTest.testAlertOnFlowFinished: Wait for deterministic flow status
About once every 200 executions on my machine this test was failing due to
the flow being killed instead of failed because jobs A and B were still in
the running state when the flow finished. Waiting for the flow to catch up
with their failures before letting job C succeed (hopefully) eliminates
this possibility.