FlowRunnerYamlTest: Wait for runner thread to exit after each test
I was seeing occasional test failures due to multiple concurrent runner
threads. In particular I was sseing a failure of testAlertOnFlowFinished
due to missing the alertOnError call, which actually did happen but only
after the test had finished. Waiting for the thread the exit before
verifying that the call occurred makes this more reliable.