Refactor time related utility methods (#2164)
PR #2113 exposed failures in tests of DefaultMailCreator class: createErrorEmail, createFirstErrorMessage and createSuccessEmail.

Failures were related to an early initialization of a static SimpleDateFormat object inside the class. As a result of this, changes to the default timezone made later in the tests did not have an effect on that object.

This PR solves this issue by eliminating the use of non-immutable static fields in the utility method that formats dates. It also moves time related utility methods from across the project into a newly created TimeUtils class.