Rename job logs link: Details -> Log (#2232)
For not-so-familiar users it's been confusing to find where the job logs are.

I don't see any reason to use "Details" as a label for the link that points to job log (the latest attempt's log).

I'm still keeping the original table header "Details". I guess that's ok.. For "flow" type nodes there is never log link currently. Maybe something else than just a log link might be added in the future to this column, IDK..

So, those Log links on the right were previously called Details.